Definite Type changes:
* Clearer more defined types were required since introduction of memory
operations

* Operations are now always of type long long
* Outputstack is now always of type double (to support memory)

* Conversions between Double -> long long uses NVIDIA intrinsics,
these do bit-wise conversions with NO rounding
 (as rounding can lead to bad memory addresses).

* Updated all examples to reflect these changes

* Fixed bug in Rule110 where first ADD_P operation had elements
 in wrong order on the output stack.

+++ examples/example2.cu
@@ -20,15 +20,15 @@ int main()
 
 
     float values[3] = {5,6,10};
-    long ops[7] = {703360,0x3CE,0x12FD,0x3CE,0x12FD,0x7E4,0x3CC};
+    long long ops[7] = {703360,0x3CE,0x12FD,0x3CE,0x12FD,0x7E4,0x3CC};
     int stack[10] = {-2,0,0,0,0,0,0,1,1,1};
     double output[6*threads*blocks] =   {0};
 
     // Allocate some memory for stack expressions
     int* stack_dev = NULL;
     int stacksize = 10;
-    long* opstack_dev = NULL;
-    long opstacksize = 7;
+    long long* opstack_dev = NULL;
+    int opstacksize = 7;
     float* valuesstack_dev = NULL;
     int valuestacksize = 3;
     double* outputstack_dev = NULL;
@@ -37,8 +37,8 @@ int main()
     cudaMalloc((void**)&stack_dev,stacksize*sizeof(int));
     cudaMemcpy(stack_dev,stack,stacksize*sizeof(int),cudaMemcpyHostToDevice);
 
-    cudaMalloc((void**)&opstack_dev,opstacksize*sizeof(long));
-    cudaMemcpy(opstack_dev,ops,opstacksize*sizeof(long),cudaMemcpyHostToDevice);
+    cudaMalloc((void**)&opstack_dev,opstacksize*sizeof(long long));
+    cudaMemcpy(opstack_dev,ops,opstacksize*sizeof(long long),cudaMemcpyHostToDevice);
 
     cudaMalloc((void**)&valuesstack_dev,valuestacksize*sizeof(float));
     cudaMemcpy(valuesstack_dev,values,valuestacksize*sizeof(float),cudaMemcpyHostToDevice);
